updated tests and ranamed methods

+++ src/main/java/com/acoustic/repository/AnnualGrossRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.acoustic.entity.AnnualGross;
 
 @Repository
-public interface AnnualGrossRepository extends JpaRepository<AnnualGross, Long> {
+public interface AnnualGrossRepository extends JpaRepository<AnnualGross, Integer> {
